pyproject.toml
[tool.poetry]
name = "dogs"
version = "0.1.0"
description = ""
authors = ["sergey-royt <goodstop687@gmail.com>"]
readme = "README.md"
packages = [{include = "dogs"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
dj-database-url = "^2.3.0"
python-dotenv = "^1.0.1"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
mypy = "^1.13.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
